requirements.txt
openai-whisper
epitran
librosa
numpy
matplotlib
rapidfuzz
flask
flask-cors
